fix: renamed file to polynomial
