+++ src/functional/base/polynomial.ts
